we now validate all input to ensure resolvable collection references
